app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
transformers>=4.35.0
torch==2.8.0
pydantic>=2.5.0
google-cloud-storage>=2.13.0
accelerate==1.11.0
psutil==7.1.3
python-multipart==0.0.20
ray[serve]==2.52.0